Add javadoc and more tests to the address registry
